Preserve prior test oracles in private modules and fix build hygiene
