+++ crates/gam-event-history/src/joint/structural_prior.rs
@@ -243,63 +243,11 @@ impl JointFunctionPriors<'_> {
 }
 
 #[cfg(test)]
-pub(super) mod tests {
+mod tests {
     use super::*;
     use crate::scalar::Mixed;
 
-    pub(in crate::joint::function_prior) fn oracle<S: JetField>(
-        function: &StructuralFunction,
-        theta: &[S],
-        rho: &S,
-    ) -> S {
-        let (log_value, log_jacobian, shape, log_gamma) = match *function {
-            StructuralFunction::CountMean { coordinate } => (
-                theta[coordinate].clone(),
-                theta[coordinate].clone(),
-                1.0,
-                0.0,
-            ),
-            StructuralFunction::TemporalVariation {
-                coordinate,
-                log_time,
-            } => {
-                let q = &theta[coordinate];
-                (
-                    add_real(&ln(&emission::softplus(q)), log_time),
-                    add_real(&emission::softplus(&q.neg()).neg(), log_time),
-                    1.0,
-                    0.0,
-                )
-            }
-            StructuralFunction::MeasurementPrecision { coordinate } => {
-                let v = theta[coordinate].scale(-2.0);
-                (v.clone(), add_real(&v, 2.0_f64.ln()), 3.0, 2.0_f64.ln())
-            }
-            StructuralFunction::InverseSoftplus {
-                coordinate,
-                log_multiplier,
-            } => {
-                let q = &theta[coordinate];
-                let s = ln(&emission::softplus(q));
-                (
-                    add_real(&s.neg(), log_multiplier),
-                    add_real(
-                        &emission::softplus(&q.neg()).neg().sub(&s.scale(2.0)),
-                        log_multiplier,
-                    ),
-                    5.0,
-                    24.0_f64.ln(),
-                )
-            }
-        };
-        add_real(
-            &rho.scale(shape)
-                .add(&log_value.scale(shape - 1.0))
-                .add(&log_jacobian)
-                .sub(&exp(&rho.add(&log_value))),
-            -log_gamma,
-        )
-    }
+    use super::super::test_support::structural_oracle as oracle;
 
     #[test]
     fn physical_shape_priors_include_exact_normalizers_and_chart_curvature() {
